+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/libs/renderer/Rectangle.kt
@@ -1,15 +1,15 @@
 package com.chattriggers.ctjs.minecraft.libs.renderer
 
 import com.chattriggers.ctjs.utils.kotlin.External
-import org.lwjgl.util.vector.Vector2f
+import javax.vecmath.Vector2d
 
 @External
 class Rectangle(
         private var color: Int,
-        private var x: Float,
-        private var y: Float,
-        private var width: Float,
-        private var height: Float) {
+        private var x: Double,
+        private var y: Double,
+        private var width: Double,
+        private var height: Double) {
 
     private var shadow = Shadow(this)
     private var outline = Outline(this)
@@ -17,35 +17,35 @@ class Rectangle(
     fun getColor(): Int = this.color
     fun setColor(color: Int) = apply { this.color = color }
 
-    fun getX(): Float = this.x
-    fun setX(x: Float) = apply { this.x = x }
+    fun getX(): Double = this.x
+    fun setX(x: Double) = apply { this.x = x }
 
-    fun getY(): Float = this.y
-    fun setY(y: Float) = apply { this.y = y }
+    fun getY(): Double = this.y
+    fun setY(y: Double) = apply { this.y = y }
 
-    fun getWidth(): Float = this.width
-    fun setWidth(width: Float) = apply { this.width = width }
+    fun getWidth(): Double = this.width
+    fun setWidth(width: Double) = apply { this.width = width }
 
-    fun getHeight(): Float = this.height
-    fun setHeight(height: Float) = apply { this.height = height }
+    fun getHeight(): Double = this.height
+    fun setHeight(height: Double) = apply { this.height = height }
 
     fun isShadow(): Boolean = this.shadow.on
     fun setShadow(shadow: Boolean) = apply { this.shadow.on = shadow }
 
-    fun getShadowOffset(): Vector2f = this.shadow.offset
-    fun getShadowOffsetX(): Float = this.shadow.offset.x
-    fun getShadowOffsetY(): Float = this.shadow.offset.y
-    fun setShadowOffset(x: Float, y: Float) = apply {
+    fun getShadowOffset(): Vector2d = this.shadow.offset
+    fun getShadowOffsetX(): Double = this.shadow.offset.x
+    fun getShadowOffsetY(): Double = this.shadow.offset.y
+    fun setShadowOffset(x: Double, y: Double) = apply {
         this.shadow.offset.x = x
         this.shadow.offset.y = y
     }
-    fun setShadowOffsetX(x: Float) = apply { this.shadow.offset.x = x }
-    fun setShadowOffsetY(y: Float) = apply { this.shadow.offset.y = y }
+    fun setShadowOffsetX(x: Double) = apply { this.shadow.offset.x = x }
+    fun setShadowOffsetY(y: Double) = apply { this.shadow.offset.y = y }
 
     fun getShadowColor(): Int = this.shadow.color
     fun setShadowColor(color: Int) = apply { this.shadow.color = color }
 
-    fun setShadow(color: Int, x: Float, y: Float) = apply {
+    fun setShadow(color: Int, x: Double, y: Double) = apply {
         setShadow(true)
         setShadowColor(color)
         setShadowOffset(x, y)
@@ -57,10 +57,10 @@ class Rectangle(
     fun getOutlineColor(): Int = this.outline.color
     fun setOutlineColor(color: Int) = apply { this.outline.color = color }
 
-    fun getThickness(): Float = this.outline.thickness
-    fun setThickness(thickness: Float) = apply { this.outline.thickness = thickness }
+    fun getThickness(): Double = this.outline.thickness
+    fun setThickness(thickness: Double) = apply { this.outline.thickness = thickness }
 
-    fun setOutline(color: Int, thickness: Float) = apply {
+    fun setOutline(color: Int, thickness: Double) = apply {
         setOutline(true)
         setOutlineColor(color)
         setThickness(thickness)
@@ -76,7 +76,7 @@ class Rectangle(
             val rect: Rectangle,
             var on: Boolean = false,
             var color: Int = 0x50000000,
-            var offset: Vector2f = Vector2f(5f, 5f)) {
+            var offset: Vector2d = Vector2d(5.0, 5.0)) {
         fun draw() {
             if (!this.on) return
             Renderer.drawRect(this.color,
@@ -98,7 +98,7 @@ class Rectangle(
             val rect: Rectangle,
             var on: Boolean = false,
             var color: Int = 0xff000000.toInt(),
-            var thickness: Float = 5f) {
+            var thickness: Double = 5.0) {
         fun draw() {
             if (!this.on) return
             Renderer.drawRect(this.color,
