+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/wrappers/objects/Particle.kt
@@ -6,34 +6,34 @@ import com.chattriggers.ctjs.utils.kotlin.MCParticle
 
 @External
 class Particle(val underlyingEntity: MCParticle) {
-    fun scale(scale: Float) {
-        this.underlyingEntity.multipleParticleScaleBy(scale)
+    fun scale(scale: Double) {
+        this.underlyingEntity.multipleParticleScaleBy(scale.toFloat())
     }
 
-    fun multiplyVelocity(multiplier: Float) {
-        this.underlyingEntity.multiplyVelocity(multiplier)
+    fun multiplyVelocity(multiplier: Double) {
+        this.underlyingEntity.multiplyVelocity(multiplier.toFloat())
     }
 
-    fun setColor(r: Float, g: Float, b: Float) {
-        this.underlyingEntity.setRBGColorF(r, g, b)
+    fun setColor(r: Double, g: Double, b: Double) {
+        this.underlyingEntity.setRBGColorF(r.toFloat(), g.toFloat(), b.toFloat())
     }
 
-    fun setColor(r: Float, g: Float, b: Float, a: Float) {
+    fun setColor(r: Double, g: Double, b: Double, a: Double) {
         setColor(r, g, b)
         setAlpha(a)
     }
 
     fun setColor(color: Long) {
-        val red = (color shr 16 and 255).toFloat() / 255.0f
-        val blue = (color shr 8 and 255).toFloat() / 255.0f
-        val green = (color and 255).toFloat() / 255.0f
-        val alpha = (color shr 24 and 255).toFloat() / 255.0f
+        val red = (color shr 16 and 255).toDouble() / 255.0
+        val blue = (color shr 8 and 255).toDouble() / 255.0
+        val green = (color and 255).toDouble() / 255.0
+        val alpha = (color shr 24 and 255).toDouble() / 255.0
 
         setColor(red, green, blue, alpha)
     }
 
-    fun setAlpha(a: Float) {
-        this.underlyingEntity.setAlphaF(a)
+    fun setAlpha(a: Double) {
+        this.underlyingEntity.setAlphaF(a.toFloat())
     }
 
     /**
